Added wait time. Fixed continuous image changing bug
Added some scripts to the game in order to implement a wait time that allows the game to wait between changing faces (1/4 of a second). Also implemented the incorrect and correct pop-ups. (now shoud display correctly and be eliminated after 1/4 of a second). There used to be a bug where clicking a button would cause the game to think that the action was a click pressed, thus if you clicked a little to long the game would go through all of the correct continuous faces that followed. I fixed it by switching ((onLeftClick)) actions to ((onLeftRelease)). Now the game only considers when the click is released so the issue has beeen solved.